Fix: Flush module prefs after a change.
This ensures that the config file remembers a 'fail to load'.  We set status to fail before loading a module, that now gets written out to the config file because we flush prefs, the dodgy module loads and crashes, and next time Audacity knows it's a fail.  If instead the module loads and does not crash on load, the original status (enabled or ask) is written back to the config file.